Database Access Package Added
Added the database access package. This will be used to get metadata
for a given table. The metadata will be used to create generated
data based on data types and the size of each field. Currently, this
is tied to Microsoft SQL Server, but other DBs could be accounted
for in the future.